python3用にreloadを削除

+++ maya/manager/_mainTemplate.py
@@ -25,7 +25,7 @@ class MainMenu(UI.MainWindowBase):
 def get_maya_main_window():
     omui.MQtUtil.mainWindow()
     ptr = omui.MQtUtil.mainWindow()
-    widget = wrapInstance(long(ptr), QWidget)
+    widget = wrapInstance(int(ptr), QWidget)
     return widget
 
 def main():
