+++ maya/manager/clickSelect.py
@@ -146,7 +146,7 @@ class CListWidgetItem(QListWidgetItem):
 def get_maya_main_window():
     omui.MQtUtil.mainWindow()
     ptr = omui.MQtUtil.mainWindow()
-    widget = wrapInstance(long(ptr), QWidget)
+    widget = wrapInstance(int(ptr), QWidget)
     return widget
 
 def main():
