+++ maya/manager/targetPosition.py
@@ -55,7 +55,7 @@ class TargetPosWindow(ui.TargetPosWindowBase):
 def get_maya_main_window():
     omui.MQtUtil.mainWindow()
     ptr = omui.MQtUtil.mainWindow()
-    widget = wrapInstance(long(ptr), QWidget)
+    widget = wrapInstance(int(ptr), QWidget)
     return widget
 
 def main():
